+++ source/cargo-verus/src/main.rs
@@ -14,76 +14,20 @@ use anyhow::Result;
 
 mod cli;
 mod metadata;
+mod plan;
 mod subcommands;
 #[cfg(any(test, feature = "integration-tests"))]
 pub mod test_utils;
 
-use crate::{
-    cli::{CargoVerusCli, VerusSubcommand},
-    subcommands::{CargoRunConfig, CargoRunPlan, NewCreationPlan},
-};
+use crate::plan::ExecutionPlan;
 
 fn main() -> Result<ExitCode> {
     use ExecutionPlan::*;
 
-    let plan = plan_execution(env::args())?;
+    let plan = plan::plan_execution(env::args())?;
 
     match &plan {
         CreateNew(creation_plan) => subcommands::create_new_project(creation_plan),
         RunCargo(cargo_run_plan) => subcommands::run_cargo(cargo_run_plan),
     }
-}
-
-pub enum ExecutionPlan {
-    CreateNew(NewCreationPlan),
-    RunCargo(CargoRunPlan),
-}
-
-pub fn plan_execution(args: impl Iterator<Item = String>) -> Result<ExecutionPlan> {
-    use ExecutionPlan::*;
-
-    let parsed_cli = CargoVerusCli::from_args(args)?;
-
-    let cfg = match parsed_cli.command {
-        VerusSubcommand::New(new_cmd) => {
-            let creation_plan = match (new_cmd.bin, new_cmd.lib) {
-                (Some(name), None) => NewCreationPlan { name, is_bin: true },
-                (None, Some(name)) => NewCreationPlan { name, is_bin: false },
-                _ => unreachable!("clap enforces exactly one of --bin/--lib"),
-            };
-            return Ok(CreateNew(creation_plan));
-        }
-        VerusSubcommand::Verify(options) => CargoRunConfig {
-            subcommand: "check",
-            options,
-            compile_primary: false,
-            verify_deps: true,
-            warn_if_nothing_verified: true,
-        },
-        VerusSubcommand::Focus(options) => CargoRunConfig {
-            subcommand: "check",
-            options,
-            compile_primary: false,
-            verify_deps: false,
-            warn_if_nothing_verified: true,
-        },
-        VerusSubcommand::Build(options) => CargoRunConfig {
-            subcommand: "build",
-            options,
-            compile_primary: true,
-            verify_deps: true,
-            warn_if_nothing_verified: false,
-        },
-        VerusSubcommand::Check(options) => CargoRunConfig {
-            subcommand: "check",
-            options,
-            compile_primary: false,
-            verify_deps: true,
-            warn_if_nothing_verified: true,
-        },
-    };
-
-    let cargo_run_plan = subcommands::plan_cargo_run(cfg)?;
-
-    Ok(RunCargo(cargo_run_plan))
 }
